+++ libraries/float16/CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.3.0] - 2024-04-17
+- Fix #12, breaking change. Thanks to Andyjbm for the measurements.
+- remove Printable interface as it makes the effective footprint larger!
+- remove getDecimals() and setDecimals().
+- patch examples and unit test for the above.
+- add example **float16_sizeof_array.ino**.
+- add **isPosInf()** and **isNegInf()**
+- add link to **float16ext** class with a larger range than float16.
+- update readme.md.
+- update unit-tests.
+
+----
+
 ## [0.2.0] - 2024-03-05
 - **warning: breaking changes!**
 - Fix #10, mantissa overflow